+++ tests/testthat/test-compare_smds.R
@@ -27,6 +27,16 @@ test_that("compare_smd", {
                       smd2 = one_boot1$df_ci$estimate[3],
                       n2 = length(datx2),
                       paired = TRUE)
+  se1 = TOSTER:::se_dz(one_boot1$df_ci$estimate[2], length(datx1))
+  se2 = TOSTER:::se_dz(one_boot1$df_ci$estimate[3], length(datx2))
+
+  one_11 = compare_smd(smd1 = one_boot1$df_ci$estimate[2],
+                      n1 = length(datx1),
+                      se1 = se1,
+                      smd2 = one_boot1$df_ci$estimate[3],
+                      n2 = length(datx2),
+                      se2 = se2,
+                      paired = TRUE)
 
   expect_equal(unname(one_1$statistic),
                unname(one_boot1$statistic))
@@ -61,6 +71,17 @@ test_that("compare_smd", {
                       smd2 = one_boot3$df_ci$estimate[3],
                       n2 = c(length(datx2),length(daty2)),
                       paired = FALSE)
+
+  se1 = TOSTER:::se_dz(one_boot1$df_ci$estimate[2], c(length(datx1),length(daty1)))
+  se2 = TOSTER:::se_dz(one_boot1$df_ci$estimate[3], c(length(datx2),length(daty2)))
+
+  one_33 = compare_smd(smd1 = one_boot3$df_ci$estimate[2],
+                      n1 = length(datx1),
+                      se1 = se1,
+                      smd2 = one_boot3$df_ci$estimate[3],
+                      n2 = length(datx2),
+                      se2 = se2,
+                      paired = TRUE)
 
   expect_equal(unname(one_3$statistic),
                unname(one_boot3$statistic))