+++ NEWS.md
@@ -3,6 +3,13 @@ NEWS
 
 **TOSTER R package and jamovi module**
 
+# TOSTER v0.6.0
+
+- Changed Glass's delta SE for paired samples (minor).
+- Added `smd_calc` and `ses_calc` for just calculating the standardized effect sizes (no tests).
+- Default CIs for for SMDs are now NCT rather than the Goulet method.
+- `compare_smd` can be supplied with user provided standard errors.
+
 # TOSTER v0.5.1
 
 - Formatting requirements for jamovi (all superficial changes)